Python/cloud_functions/Auth_Func/src/requirements.txt
google-cloud-secret-manager==2.12.6
python-dotenv==0.21.0
Pyrebase4==4.5.0